Move legacy routing class and alias
